requirements.txt
dpath==2.1.3
pymediainfo==6.0.1
PyYAML==6.0
